app/dma-perf: fix mbuf IOVA with remote address
Function call rte_mbuf_iova_set() on architecture using IOVA address
as VA would not set iova address in buf_iova field of mbuf. For DMA
transfer from "mem to dev" and "dev to mem", buf_iova of mbuf must be
overwritten with the user configured remote address which as part of
this change set is done using external buffer attach/detach API.

Fixes: 56f7edf9d3a3 ("app/dma-perf: add PCI device support")

+++ app/test-dma-perf/main.h
@@ -63,7 +63,7 @@ struct test_configure {
 	uint8_t dcoreid;
 	uint8_t pfid;
 	uint16_t vfid;
-	uint64_t raddr;
+	uintptr_t raddr;
 };
 
 void mem_copy_benchmark(struct test_configure *cfg, bool is_dma);
